Fix multi-user temp status bug in vis chatbot
This bug occurs when the previous user exits leaving the inferencer still
working on his/her request, the remaining response will be left in the queue
and interferes with later users. Add mechanisms so that a user only starts
his/her request AFTER the request_queue and response_queue is emptied.

- Also, add demo.queue(api_open=False) so no one can bypass the queue